nios2: Added breakpoint exception handler
This is an initial attempt of enabling gdb/gdbserver application debug support
on the nios2 platform. To distinguish a gdb breakpoint from a system call, we put
an immediate value of 1 into the trap instruction. In the instruction set the immediate
value is undefined for trap, but apparently the instruction decoder recognizes the
instruction anyway. Thus we in the kernel exception handler can do an additional check
for imme=1 if a trap instruction is detected, and then run an appropriate handler.

Signed-off-by: Atle Nissestad <atle@nissestad.no>