nios2: fix clobbered register in local_irq_restore inline asm
gregor1 identified a bug and submitted a patch.

The function local_irq_restore(x) in include/asm/system.h in the current
linux kernel seems to have a big problem. The register r8 will not be
reserved. So the compiler can use this variable for the normal c-code.

This patch adds register r8 to the clobber list of the inline asm.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>