open_eth: use basic mii regs and uncached buffer
This patch added basic mii regs support so that almost any phy chips can
be used. However, mii irq is not supported.

The patch changed the preallocated tx rx buffer to uncached, so that
cache flushing will not be needed.

FIXME: This mii hack should be ported to use the generic phylib. The
phylib can search mii address during initialization. The mii interrupt
can be handled.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>