open_eth: revert to use cached buffer
The performance of cached and uncached buffer are very close. So it is
reverted to use cached buffer.

The irq disable to protect xmit queue head is changed to bottom half
disable, because we are using bottom half (tasklet/softirq) to handle tx
interrupt.

Using the Avalon Opencore 10/100 MAC update from Jake, it handles
unaligned buffers and can work without XXBUFF_PREALLOC.

FIXME: This driver is not SMP-safe.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>