nios2: dont use uncached load store instructions
The uncached memory access can be achieved using ioremap addressing.
It will be more consistent with MMU usage. On the other hand, using
special uncached load store instructions only make it confusing, and the
usage is deprecated here.

The io.h header file was synced to that of m68knommu. And io.c was
included as inline.

The memcpy_fromio was changed to byte copy only version, as the memcpy
was optimized with word copy which might have problems on 8 bits or 16
bits memory devices.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>