drivers/ata/pata_altera_cf: New PATA driver for Altera CompactFlash.
This patch adds a new PATA driver to support the Altera SOPC Builder
CompactFlash component as a PATA device with support for insertion
and removal of CF cards.  It uses the platform driver model.

For each probed platform device, the driver spawns a kthread to
handle insertion and removal of CF cards by creating and removing the
underlying ATA host interface.  The Altera CompactFlash component
includes a register bit to indicate whether a CF card is detected
and an interrupt triggered when this bit changes.  The driver has a
handler for this interrupt which flags that something has changed and
wakes up the kthread to deal with it.

The platform device uses 2 IORESOURCE_MEM resources and 2
IORESOURCE_IRQ resources.  The first memory resource is for the
CompactFlash component's IDE registers and the second is for the
component's CF status and control registers.  The first IRQ resource
is for the IDE and the second is for the CF detection interrupt.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>