nios2: support pata_altera_cf.
This patch adds platform device support for my pata_altera_cf driver.

If the SOPC system includes an Altera CompactFlash component, and the
the kernel configured to build the pata_altera_cf driver, a platform
device is created for this driver.  If the pata_altera_cf driver is
not configured, but pata_platform is, a platform device is created for
that driver instead.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>