nios2: If not optimizing for size, unroll the string I/O loops.
Use Duff's Device to unroll the loops in the nios2 implementations
of the string I/O functions (such as ioread8_rep()).  This can speed
up the string I/O functions by a few percent, depending on how long
an individual I/O operation takes.

Only do this if not optimizing for size.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>