nios2: Display CPU clocking in /proc/cpuinfo if available
Newer generated header provide CPU_FREQ which can be used in
show_cpuinfo() to get the CPU frequency. Though default_mmu.h currently
shipped with nios2mmu linux does not include it, so we need to fake it
if it is not defined (setting it to 0 should make it obvious to
everyone).

Signed-off-by: Tobias Klauser <klto@zhaw.ch>