altera_tse: Simplify MDIO access
MDIO register access for altera_tse was implemented in an own platform
driver which claimed the TSE_RESOURCE_MAC_DEV resource which was also
claimed by the altera_tse platform driver itself. This caused several
problem, so simplify the MDIO register access by moving the struct
mii_bus allocation and setup to the probe function of the altera_tse
platform driver after the resource is claimed there. The MDIO read/write
functions are also moved to the altera_tse driver so we can get rid of
the alter_tse_mdio module alltogether.

Signed-off-by: Tobias Klauser <klto@zhaw.ch>