altremote: Correct watchdog status and watchdog counter
The current "watchdog enabled" and "watchdog timeout" values should be
read from the "Past Status 1" set of registers in application mode, not
from the "Current Status" set of registers.  These values read back as 0
in factory mode, but don't apply there. Also, the watchdog timeout value
iread back in application mode should be shifted right by 17 bits
because the 12-bit value in bits 28 to 17 matches the 12-bit value
written to the watchdog timeout register in factory mode.

(Note: we're assuming it's the Cyclone III type of remote update
component here, because that's all that's been implemented by this
driver so far.)

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>