altremote: Show configuration mode and previous boot address in status
When showing the status, report whether it's in factory mode or
application mode, and report the previous boot address if available.
Report the previous boot address as 0 if not available.

Note: the previous boot address is only available in factory mode, but
not if it's the initial configuration from power up.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>