nios2: Fix bug in local_tlb_flush_all which corrupted tlbmisc register
I believe I have tracked down a bug in the TLB code which caused
problems during rapid forks.  Basically, the following shell script
would fail reliably:

#!/bin/sh
while true; do
    if [ `ps w | grep -v grep | grep -E 'nfsiod|ps' | wc -l` == "2" ]
        then
          /bin/echo -n "."
        else
           echo "*** FAILURE ***"; exit 1;
    fi
done

The kernel code which flushes the TLB's did not protect one of the TLB
registers (tlbmisc) and this caused subtle corruption of the hardware
TLB entries.  It seems to only happen when a process forks and it's
pretty quickly "corrected" by subsequent tlb replacement so the bug is
intermittent and shows up as 3-4 different types of userland faults.

A fix follows which saves the original value of the tlbmisc register
during the clear.

The patch also includes a fix to one of the tlb clearing routines which
was not correctly setting the "write" bit.  This bug is probably benign,
because the process pid's would never match, but may slightly improve
performance.

(to find this was tough; I modified my nios emulator to record all the
userland reads and writes with their associated tlb entries and then I
could see that something wasn't correct - every once and a while a
process would write with one tlb and read the same address with another
(incorrect) tlb and then fault, since it would read back bad data.  It
looks like this bug has been there since the initial port.)

simulation is my friend :-)

Signed-off-by: Brad Parker <brad@heeltoe.com>
Signed-off-by: Tobias Klauser <klto@zhaw.ch>