of/flattree: Add Kconfig for EARLY_FLATTREE
Follow commit e6ce1324e4f08b0d984340201a125806dae0e9a6,

The device tree code is now in two pieces: some which can be used generically
on any platform which selects CONFIG_OF_FLATTREE, and some early which is used
at boot time on only a few architectures.  This patch segregates the early
code so that only those architectures which care about it need compile it.
This also means that some of the requirements in the early code (such as
a cmd_line variable) that most architectures (e.g. X86) don't provide
can be ignored.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>