fdt: add early address translation
This patch adds address translation when working with a flattened tree.
This patch is also send to the device-tree mailinglist and I hope I will
get it aqccepted there but I'm pushing it here as well since we really
need this to get early_printk support on systems where the uart used as
early console is attached to a bridge. One example of such a design is
the neek_ocm_spi_mmu example design.

Signed-off-by: Walter Goossens <waltergoossens@home.nl>