nios2: Removed unused code
This code was ifdeffed by CONFIG_CRC_CHECK but CRC_CHECK is not an option in
Kconfig and CONFIG_CRC_CHECK is defined nowhere. Further more "na_flash" is
used in this code which leads me to believe that it might have been used in
the nommu code. This code has been in git since the initial import but I
seriously doubt whether or not is has been ever used. It might have been a
leftover from the nios (note the missing 2 or II) days.

Signed-off-by: Walter Goossens <waltergoossens@home.nl>