nios2: Check DIV/MUL/MULX config at runtime
When using dts we are no longer able to determine mismatches between CPU
support and kernel configuration for the DIV/MUL/MULX instructions at
compile time. So change it to a runtime check in cpuinfo and remove the
compile time check.

This also changes the Kconfig variables to becom integers, otherwise we
couldn't compare them to the fields in cpuinfo.

Signed-off-by: Tobias Klauser <klto@zhaw.ch>