nios2: Define region base addresses in Kconfig
The region base addresses are (usually) always the same and do not need
to come from the dts (or the generated headers) so define them in
Kconfig and allow the user to override them if needed (can't think of a
use case where someone would want to do this, but still).

All region base addresses except IO_REGION_BASE are 0 on NOMMU. This
also allows us to eliminate some '#ifdef CONFIG_MMU' in the code.

Signed-off-by: Tobias Klauser <klto@zhaw.ch>