nios2: Use generic arch_get_unmapped_area
arch_get_unmapped_area in arch/nios2/mm/mmap.c is an exact copy of the
MIPS code and does not seem to carry any nios specifics.

As SHMLBA is equal to PAGE_SIZE on nios2, the COLOR_ALIGN macro is
equivalent to PAGE_ALIGN and thus this function is equivalent to the
generic arch_get_unmapped_area in mm/mmap.c.

Signed-off-by: Tobias Klauser <klto@zhaw.ch>