nios2: Get rid of asm/nios.h
Allow the memory base address and memory size to be defined using
Kconfig. Thus we no longer need to fiddle with asm/nios.h to adapt it to
the actual memory values of the system.

Also make the cache settings for the early boot code tweakable via
Kconfig (but these normally shouldn't be touched).

Signed-off-by: Frédéric LAMBERT <frdrc66@gmail.com>
Signed-off-by: Walter Goossens <waltergoossens@home.nl>
Signed-off-by: Tobias Klauser <klto@zhaw.ch>