altera_tse: Generally use MDIO space 1
The ip core has two mapping possibilities to access the PHY registers
(MDIO space 0 and 1). However, if the ip core is built for SGMII
MDIO space 0 is occupied by the PCS interface registers.

Since we support one PHY, only, we should take MDIO space 1, because this
is the mapping possibility which is always present - no matter if the ip
core is built for SGMII or any other mode.

Signed-off-by: Joachim Foerster <joachim.foerster@missinglinkelectronics.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>