altera_tse: Fix race condition in OBHW check
Fix a possible race condition in the OBHW (Owned By Hardware) check
loop in the TX part of the tse_poll function. The SGDMA could update
the OBHW bit after the check is done and the check if sgdma_async_write
should be called.

Without this fix there is a possibility that outgoing packet could get
stuck, and would not be sent until there was an incoming packet.

Signed-off-by: Johan Karlsson <johan.karlsson@xelmo.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>