nios2: Reserve bootmem for chosen DTB, if within main memory and not built-in
A passed DTB via bootloader or a DTB at a hard-coded address might not
always generally be within (current code: passed DTB) or
out of main memory (current code: hard-coded address).
For example a DTB could be passed via a small in-FPGA memory.

Note: If the DTB is built-in, base >= memory_start is still true, since
memory_start isn't the actual start of main memory. It is the first page
after the kernel binary.

Signed-off-by: Joachim Foerster <joachim.foerster@missinglinkelectronics.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>