nios2: Remove code duplication in cmd_line assignment/filling.
If no command line is passed via a boot loader, cmd_line is already
assigned in setup.c:nios2_boot_init() with exactly the contents
of CONFIG_CMDLINE. Even if we don't compile with CONFIG_PASS_CMDLINE,
this assignment is done.
So we certainly don't need this memcpy() to cmd_line here in
setup_arch().

Signed-off-by: Joachim Foerster <joachim.foerster@missinglinkelectronics.com>