nios2: Fix __access_ok macro semantics for NOMMU
As currently we (incorrectly) use memory_start as the first address
after the kernel in memory (instead of the first address of memory), we
need to check against CONFIG_MEM_BASE (the actual base address of kernel
memory) in __access_ok (we just need to check wheter the address is in
physical memory as there are no separate kernel and user address spaces).

Also the last address at memory_end isn't a valid address, so fix this off-by-one
error too.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>