nios2: Make ELF_NGREG large enough to hold all coredumped registers
In ELF_CORE_COPY_REGS, 33 registers (sizeof(struct pt_regs) +
sizeof(struct switch_stack)) are copied to the buffer, but the buffer
was only allocated to hold ELF_NGREG (sizeof(struct pt_regs)) registers.

This resulted in a buffer overflow when writing the corefile and a
corefile which didn't contain the .reg section was produced. This in
turn lead to the following message when loading the corefile into GDB:

  warning: Couldn't find general-purpose registers in core file.
  #0  0x00000000 in ?? ()

Extending the ELF_NGREG fixes this behaviour and the resulting corefile
can sucessfully be used within GDB.

Reported-by: Yuriy Kozlov <ykozlov@ptcusa.com>
Reference: http://www.alteraforum.com/forum/showthread.php?t=21237
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>