nios2: Force build of compiled-in device tree
If CONFIG_DTB_SOURCE is changed but the dts has an older (or the same)
timestamp as the previously configured, the dts will not get recompiled.

Until we find a more elegant solution to only recompile the dts in case
CONFIG_DTB_SOURCE got changed, we should just force the dts being
compiled everytime.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>