nios2: Add custom byteswap instruction support
Altera provides an endian converter custom instructions which can
be enabled in SOPC builder.  It reverses the bytes in a 32-bit word
in one cycle.  Use this instruction for 16 and 32-bit byteswap macros
and add a Kconfig option to enable it.

This reduces swab16 (a.k.a. ntohs/htons) from 5 instructions to 3 and
swab32 to one instruction.

Signed-off-by: Yuriy Kozlov <ykozlov@ptcusa.com>
[tklauser: Add Kconfig var for opcode number and other slight adaptions]
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>