net: ethernet: altera_tse: Set random MAC address if none available
Up to now we needed a MAC address from the device tree. Drop this
requirement and generate a random MAC address (like other drivers to
this) if no valid address is passed via device tree (e.g. also an
invalid address like 00:00:00:00:00:00 from the device tree will be
considered invalid).

Also set the MAC address to the address register in the probe function.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>