gpio: gpio-altera: Add proper GPIO IRQ support
As only one IRQ is assigned to the whole GPIO controller, we need to
handle it in the driver and determine which GPIO caused the interrupt.
Also the edgecapture register will need to be cleared (at least in
certain configurations of the controller, it won't harm in the other
cases).

Thus setup an irq handler in the driver and assign a virtual IRQ number
to each GPIO pin. altera_gpio_irq_handler() will then determine which
pin(s) caused the interrupt and call the according "real" interrupt
handler.

This depends on commit c6d91791 (nios2: Allow for additional 32 IRQs) to
run properly.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>