net: ethernet: altera_tse: Omit padding of frame to multiple of 4
We only need to align the start of the frame (and even that could be
omitted in the case where the SGDMAs and the MAC allow unaligned
transfers) but not the length.

Increasing the length actually caused some problems if an MTU is set on
the device which is not a multiple of 4. Frames then got padded beyond
the MTU and could potentially get dropped by other hosts in the network.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>