Revert "spi_altera: Allow to explicitely override bus number via dts"
This reverts commit f2c9356fc8ea82cea920396407a7f07ea44c501d.

It is illegal for device drivers to modify the pdev->id field.  Once a
device is registered, the id and name cannot be changed.

With 07d57a32fb6eb2da017796e038682f817a4f685e "drivercore: Output common
devicetree information in uevent" in place we now have a better way to
resolve the /sys/devices/ path for a specific devicetree node.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>