nios2:  don't reimplement force_sigsegv()
Follow de8f2ddd.

Instead of open coding the sequence from force_sigsegv() just call it.
This also fixes a bug because we were modifying ka->sa.sa_handler (which
is a copy of sighand->action[]), whereas the intention of the code was to
modify sighand->action[] directly.

As the original code was working with a copy it had no effect on signal
delivery.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>