nios2: correct umount syscall
The syscall table for nios2 currently points to the sys_umount() for
both the __NR_umount (==22) and __NR_umount2 (==52) syscall numbers.
This results in sys_umount() being called with a garbage 'flags'
parameter for the __NR_umount syscall, so it is likely to return
-EINVAL.

The syscall table entry for __NR_umount should point to sys_oldumount().
Change it.  Also remove a misleading comment for the syscall table entry
for __NR_umount2.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>