arch/nios2: (!CONFIG_MMU) fix cache_invalidate_inst()
arch/nios2/mm/cacheflush-nommu.c contains a static inline function
cache_invalidate_inst() called by cache_push() and icache_push().
Currently, when the length to be flushed exceeds the size of the
instruction cache, it rounds the start and end of the desired area
incorrectly, using the cache size instead of the cache line size, so it
is unlikely to flush all the desired area.  An extra cache line was
added to the end of the region to be flushed, possibly as a partial
workaround for the real bug.

Round the start and end of the desired area correctly and don't flush an
extra cache line at the end.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>