arch/nios2: (!CONFIG_MMU) Avoid cache write-back for DMA_FROM_DEVICE
Change __dma_sync() to use nios2_clear_dcache_range() for the
DMA_FROM_DEVICE direction.  This invalidates cache lines without doing a
write-back of dirty cache lines first (except for the start and/or end
of the range if they are not aligned to a cache-line boundary).

This has only been implemented for !CONFIG_MMU so far, so for CONFIG_MMU
it still uses the regular flush_cache_range() instead.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>