arch/nios2: (!CONFIG_MMU): simplify cache_invalidate_inst()
arch/nios2/mm/cacheflush-nommu.c contains a static inline function
cache_invalidate_inst() called by cache_push() and icache_push().
Currently, it adds on a full cache-line to the end of the region to be
invalidated, but it only needs to round the end up to a cache-line
boundary.  Also, since the NiosII "flushi" instruction ignores the 'tag'
part of the address, there is no need to zero out the tag part of the
`paddr` variable value when calculating the values for the `sset` and
`eset` variables.  This simplifies the expressions used to set those
variables a bit.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>